Move devnet config from `src` to `tests` directory
